Add new oak_io crate
Added a new oak_io crate for shared structs used by the runtime and SDK.

Moved the ReadHandle and WriteHandle structs to the new crate.

Updated the Runtime to use ReadHandle and WriteHandle.
