+++ examples/Cargo.toml
@@ -38,6 +38,7 @@ members = [
 oak = { path = "../sdk/rust/oak" }
 oak_abi = { path = "../oak_abi" }
 oak_derive = { path = "../sdk/rust/oak_derive" }
+oak_io = { path = "../oak_io" }
 oak_runtime = { path = "../oak/server/rust/oak_runtime" }
 oak_tests = { path = "../sdk/rust/oak_tests" }
 oak_utils = { path = "../oak_utils" }
