+++ oak_io/Cargo.toml
@@ -1,0 +1,11 @@
+[workspace]
+
+[package]
+name = "oak_io"
+version = "0.1.0"
+authors = ["Conrad Grobler <grobler@google.com>"]
+edition = "2018"
+license = "Apache-2.0"
+
+[dependencies]
+oak_abi = { path = "../oak_abi" }
