+++ oak_io/src/handle.rs
@@ -1,0 +1,45 @@
+//
+// Copyright 2020 The Project Oak Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+use oak_abi::Handle;
+
+/// Wrapper for a handle to the read half of a channel.
+///
+/// For use when the underlying [`Handle`] is known to be for a receive half.
+#[derive(Copy, Clone, PartialEq)]
+pub struct ReadHandle {
+    pub handle: Handle,
+}
+
+impl std::fmt::Debug for ReadHandle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "ReadHandle({})", self.handle)
+    }
+}
+
+/// Wrapper for a handle to the send half of a channel.
+///
+/// For use when the underlying [`Handle`] is known to be for a send half.
+#[derive(Copy, Clone, PartialEq)]
+pub struct WriteHandle {
+    pub handle: Handle,
+}
+
+impl std::fmt::Debug for WriteHandle {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> Result<(), std::fmt::Error> {
+        write!(f, "WriteHandle({})", self.handle)
+    }
+}
