+++ oak_io/src/lib.rs
@@ -1,0 +1,21 @@
+//
+// Copyright 2020 The Project Oak Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+
+//! Shared data structures and functionality for inter-node communication.
+
+pub mod handle;
+// TODO(#1324): Move shared structs from the Runtime and SDK (like `Sender`, `Receiver`,
+// `Encodable`, `Decodable` and `Message`/`NodeMessage`) here as well.
