+++ sdk/Cargo.toml
@@ -17,6 +17,7 @@ oak = { path = "rust/oak" }
 oak_abi = { path = "../oak_abi" }
 oak_app_build = { path = "rust/oak_app_build" }
 oak_derive = { path = "rust/oak_derive" }
+oak_io = { path = "../oak_io" }
 oak_runtime = { path = "../oak/server/rust/oak_runtime" }
 oak_tests = { path = "rust/oak_tests" }
 oak_utils = { path = "../oak_utils" }
