+++ sdk/rust/oak/Cargo.toml
@@ -8,6 +8,7 @@ license = "Apache-2.0"
 [dependencies]
 oak_abi = "=0.1.0"
 oak_derive = "=0.1.0"
+oak_io = "=0.1.0"
 prost = "*"
 prost-types = "*"
 fmt = "*"
